+++ Code/Engine/CoreUtils/Image/Formats/TgaFileFormat.h
@@ -8,5 +8,6 @@ public:
   virtual ezResult ReadImage(ezStreamReaderBase& stream, ezImage& image, ezLogInterface* pLog) const EZ_OVERRIDE;
   virtual ezResult WriteImage(ezStreamWriterBase& stream, const ezImage& image, ezLogInterface* pLog) const EZ_OVERRIDE;
 
-  virtual bool IsKnownExtension(const char* szExtension) const EZ_OVERRIDE;
+  virtual bool CanReadFileType (const char* szExtension) const EZ_OVERRIDE;
+  virtual bool CanWriteFileType(const char* szExtension) const EZ_OVERRIDE;
 };