+++ Documentation/WhatsNew.md
@@ -8,6 +8,7 @@ Milestone 5
   * Added ezJSONReader which generates a high level representation of a JSON document, that can be used to lookup data easily.
   * Added ezExtendedJSONReader which generates a high level representation of a JSON document, but also uses the extended JSON syntax for more precise type information and values.
   * Added ezImage to the CoreUtils library, which allows to easily read / convert / write images of different file types.
+  * Supported image file types: BMP, DDS, TGA
 
 Milestone 4
 -----------